fixed level 3 warnings, cleaning libressl and fixing libressl test infra

+++ contrib/win32/win32compat/pwd.c
@@ -171,9 +171,9 @@ get_passwd(const char *user_utf8, LPWSTR user_sid)
 		goto done;
 	}
 
-	uname_upn_len = strlen(uname_utf8) + 1;
+	uname_upn_len = (DWORD) strlen(uname_utf8) + 1;
 	if (udom_utf8)
-		uname_upn_len += strlen(udom_utf8) + 1;
+		uname_upn_len += (DWORD)strlen(udom_utf8) + 1;
 
 	if ((uname_upn = malloc(uname_upn_len)) == NULL) {
 		errno = ENOMEM;
